% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot data on a map of a state for a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number.}

\item{year}{Year.}
}
\description{
This function plots a map indicating the points where the accidents have been reported.
If the state number provided is not in the data, it outputs an error. Similarly if the year is not present
or there is no data to plot, the function terminates with a message "no accidents to plot"
}
\examples{
fars_map_state(6,2013) creates a map plot for state number 6 for the year 2013
}
