% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Function to read a CSV file and create a data frame.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Complete path to a CSV file.}
}
\value{
Data Frame
}
\description{
This function reads a csv file and creates a data frame.
This function requires packages - deplyr and readr.
If the file name is not present it errors out with a message
}
\examples{
dat <- fars_read("data/accident_2013.csv")
}
