% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read CSV file for a given year  and build a data frame}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A data frame (with columns MONTH and Year) for each of the years.
}
\description{
Reads the appropriate .csv.bz2 file for a given year and builds a data frame with two columns - Month and Year selected from the CSV.
If the year is invalid, it outpus a warning saying "invalid year" and does not produce any output
}
\examples{
fars_read_years(c(2013,2014,2015)) will create data frames for each of the years
}
