% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of accidents by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
summary data frame of number of accidents per month for a given vector of years
}
\description{
This function returns a summary data frame of number of accidents per month for a given vector of years.
}
\examples{
fars_summarize_years(c(2013,2014,2015)) creates a data frame with the columns - MONTH, `2013`, `2014`, and `2015`
}
