% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Function to return a filename for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The 4 digit year for the filename to be created.}
}
\value{
File name
}
\description{
This function returns an appropriate filename with ".csv.bz2" extension.
}
\examples{
make_filename(2013) outputs the filename as [1] "accident_2013.csv.bz2"
}
