% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverser.R
\name{sheilarev}
\alias{sheilarev}
\title{This function reverses a string or numbers}
\usage{
sheilarev(she)
}
\arguments{
\item{she}{A string or number}
}
\value{
The reverse of the provided string or number
}
\description{
This function reverses a string or numbers
}
\examples{
sheilarev(1234)
sheilarev("goo")
}
